backend tested without RABBIT

+++ src/main/java/customer_complaint/customer_complaint/service/impl/CategoryServiceImpl.java
@@ -35,6 +35,9 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void delete(Long id) {
+        if (!categoryRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Category not found");
+        }
         categoryRepository.deleteById(id);
     }
